requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
sqlalchemy>=2.0.0
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pytesseract>=0.3.10
Pillow>=9.5.0
pdf2image>=1.16.0
python-dateutil>=2.8.0
psutil>=5.9.0 